feat: loading environment variables
