+++ src/index.ts
@@ -1,11 +1,17 @@
 import 'reflect-metadata';
-import { LocalDB } from './data/connection/local.db';
-import { createServer } from 'http';
-import { App } from './app';
+import { Connection } from './data/config/connection';
+import { Server } from './server';
+import dotenv from 'dotenv';
+
+// Config environments
+dotenv.config();
 
 // Database connection
-LocalDB.then(() => console.log('🟢 Database connected')).catch(() => console.log('🔴 Database not connected!'));
-
-// Server listening
-const httpServer = createServer(App);
-httpServer.listen({ port: 3000 }, () => console.log(`🟢 GraphQL is running on http://localhost:3000/graphql`));
+Connection().then(() => {
+  console.log(`🟢 Database connected`);
+  Server().then(() => {
+    console.log(
+      `🟢 GraphQL is running on http://localhost:${process.env.PORT}${process.env.GRAPHQL_PLAYGROUND_ENDPOINT}`,
+    );
+  });
+});
